FX/Phase2/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx>=0.27.0
python-dotenv==1.0.0
pydantic==2.5.2

# VTT parsing
webvtt-py>=0.5.0

# Google API
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0

# LLM (Optional)
openai>=1.35.10

# Audio transcription (Optional - requires ffmpeg installed on system)
# Uncomment if using Whisper for audio transcription
# openai-whisper==20231117
# ffmpeg-python==0.2.0